Change the default behaviour of the Roi containment for offline reference tracks
The monitoring was changed to use a new selection of refernec tracks that were strictly contained
within an Roi rather than contained within the Roi *and* within the more restrictive eta limits.

This causes issues with the comparison with previous versions so the default behaviour has been
changed to no longer use the rigorous containment, and it can be enabled with a flag and job option.

In addition, some extra flags have been added to the art run scripts to aid local running and
debugging. A new local utility to help the update of the art reference files has also been added


Former-commit-id: 09ad875fe3002fdc6e1e0d20edc223dd6c607adb