Migrate CscRawDataMonitoring to use ReadHandles for data access
Most of the changes are straightforward, simply switching data access from direct retrieval from the eventStore or the activeStore to using a ReadHandle.
Changes to CSCSegmValAlg are slightly more significant, as it was set up to be able to process multiple segment collections.
This was a relic of the Run-1 situation, in which there were two segment collections, one for each reco algorithm, but now there is only one segment collection, making this unnecessary.
It seemed pointless to create a ReadHandleArray for just the one segment collection, so I modified the code so that it needs to deal with only segment collection.
This mainly involved switching from vectors of histograms and MonGroup objects to single objects, and changing the configuration to reflect that it is only processing a single collection.
Mentioning ATLASRECTS-4390 for bookkeeping purposes.


Former-commit-id: ba015adc23a2f0ea84a4be6c7fdc62f05183893c