Update TPCnv unit tests using HepMcParticleLinkCnv_p2 to use Athena::LeakCheck
By creating a dummy `HepMcParticleLink` ahead of creating the `Athena::LeakCheck` we can
avoid leak warnings from the static variable created by the class. The same approach is
used to mask a similar feature in the DVL Info. This all means that we can add back the
leak checking to these unit tests. In the longer term we will try to remove the static
variables from the `HepMcParticleLink` implementation - see ATLASSIM-3605.


Former-commit-id: 0b8192c6cb65152d3dbab6a63dbeca8172802e57