extract common geometry config from InDetGMJobProperties
The common geometry configuration stored in the Oracle table
"AtlasCommon" can only be accessed via InDetGMJobProperties as it
is implemented as properties of GeometryFlags which is member of
InDetGMFlags. Because the properties are (mostly) general (like
the run period) they should not be part of the ID only, but really
"common". It is already bad enough that the only specific field
in the table is called "STRIPGEOTYPE" ... On top, it turned out
that InDetGMJobProperties.py and LArGMJobProperties.py are quite
carbon copies of each other. These flaws are addressed in the
implementation of CommonGMJobProperties.py which provides access
to the common properties. Secondly it serves as base-class for the
classes in InDetGMJobProperties.py/LArGMJobProperties.py to reduce
code duplication. The modification is purely technical and should
not change anything. It passes the Tier0 tests. All clients should
be adjusted, but it was not possible to find test jobs for all of
them.


Former-commit-id: c79918fa8581545ffea4c3f5519b589f111c7d6c