ISF_FatrasToolsG4: Fix for PDGToG4Conv tool being initialized too early.
With tool auto-initialization, the PDGToG4Conv tool in G4ParticleDecayHelper
was getting initialized too early --- Geant will eventually throw an
exception if this happens before the run manager is set up.
Disable the tool handle in G4ParticleDecayHelper::initialize()
to prevent this.

Needed to fix ATN failures in ISF_Validation.



Former-commit-id: 0fec474b6294a14526f942f16f7cd5a9a47f5787