Make TruthStrategies know the detector regions where they should be applied
This commit extends the `ITruthStrategy` interface adding an
`appliesToRegion(AtlasDetDescr::AtlasRegion geoID)` method which
returns `true` if the truth strategy should be applied to the
detector region represented by `geoID`.
The implementation classes now have a new member variable which
is the list of detector regions where they are valid.
This code is all the same between the implmentations so may argue
for having an implementation base-class in the future.
This is part of the changes for ATLASSIM-3377.


Former-commit-id: 7320bd2b815f48e9433d94ea67eabfa6a48bf86f