Blocking use of AllVariables for 1-2% DAOD formats and dropping HIGG2D2
This MR extends the list of formats that are blocked from using the AllVariables method of the slimming helper to those formats which are 1-2% of the size of their input AOD. Once accepted, all physics DAODs > 1% of their input will be blocked form using AllVariables. The obsolete HIGG2D2 is also dropped from ART and the list of flags.


Former-commit-id: 11c48f1697f7672d9dd7e94dafff14c40fe6a11a