Adding pre-selection sequencers to flavour tagging formats
This MR introduces new pre-selection sequencers, to which a kernel running the skimming tools is attached. The main sequence is then attached to the new sequencer. In this way we run the jet and flavour tagging only for events that are going to saved. Significant CPU reductions are achieved in this way.


Former-commit-id: 6c302d4cafb1c1aa1dff3b562e9a27b83a7a0b4e