This merge request implements proper combinations of 2017/2018 algorithms for tau triggers (ATR-17842). We can now easily define chains with MVA TES, BDT track counting, and RNN ID, switched on/off. It also fixes a couple of inconsistencies with previous chains that went into the menu (ATR-17807): - for tracktwomvaTES chains (MVA TES + precision track counting), MVA TES was applied at preselection step, but calo TES was used at precision step - RNN_tracktwo chains using RNN ID but 2017-like FTF track preselection were not scheduling RNN ID but BDT ID (the whole 2017 precision sequence, actually).
For now, 2017 and 2018 precision sequence were kept separate for safety, but the new scheme could have been used for 2017 chains.

Also, better memory usage is expected, by using the same tools in TauRecMerged_TauPrecisionMVA and TrigTauRecMerged_TauPrecision.
As long as the tool configuration is identical in the different sequences, it should be OK to use the same tools, instead of re-declaring new tools as done before (by using different tool names).
Here also, a large reduction in the number of tools can be easily achieved, presumably.
Because all TrigTauRecMerged classes declare independent sets of tools, while a large fraction can be shared between sequences.
This could be tested and implemented in a follow-up MR.

However, the fact that we need different sequences, for different combinations of 2017/2018 tools, cannot be really optimised, as far as I understand.


Former-commit-id: 5828808bd94f3ccca3aeb94b7e67552d51d786c1