Fix the reference files for the new pu60 job
Fix the reference histograms in the plotting tests for the new pu60 test.
Also takle the oportunity to tidy up the ps listing output in the template.sh job

In principle all the ART run scripts should be recompiled with this, but that would
increase the number of changes in this merge request so that can easily wait until
all the jobs need to be modified the next time.


Former-commit-id: eff4fe19499c1155b5a0f8ce945a9facb55c406d