Resubmitting MR 9783 to 21.1
Hello, this MR is a resubmission of MR 9783, with identical content (except trivial fix).
The goal is to target 21.1 instead of 21.0 with 'alsoTargetting 21.1'.
Hopefuly this will speed up the merging, and also compilation had failed in MR 9783, due to something unrelated, which may need manual intervention on the CI side.
Just trying to gain time, as this is urgently needed for validating tau triggers in HLT reprocessing.
Thanks!


Former-commit-id: b04c7016120ea419593793e9e8250d2a6dbfe2b7