SGTools: Fix potential dangling proxy pointer.
If a proxy was registered using only addToStore such that the first call
was for a CLID different than proxy->clID(), then the primary_sgkey
recorded is based on the CLID used for this first record, not
proxy->clID().  [This ordering used to happen when ProxyProviderSvc
initialized a proxy.]  This in turn would confuse removeProxy when a proxy
was to be actually deleted, causing a dangling reference to the deleted
proxy to remain in m_keyMap.  This could then lead to a crash in overwrite().
This was observed in a derivation workflow.  Fix this.



Former-commit-id: b7782968d9509c22152345c0b8357b642507fa4d