manually sweeping !9580 from master to 21.3
For further details on the topic "extract common geometry config
from InDetGMJobProperties" please see the original merge request.
The manual sweep was a bit of a pain because in 21.3 there is the
additional property "useDynamicAlignFolders()" of the former
GeometryFlags still present. Therefore a few more files needed
to be modified (hopefully i caught all of them). In any case,
running the frozen Tier0 test passed, i.e. there should be no
immediate failure. Nevertheless it would be a good idea to monitor
the runtime tests more closely after the merge to see if sth
crashes.


Former-commit-id: 8bf0b675ed673b4bf5e61f33e7d455203f46921e