Migrate RpcRawDataMonitoring to use ReadHandles for data access
Mostly this just changes direct retrievals from the store to access via ReadHandles instead.
The main additional change is that RpcRawDataMonitoring now has an m_isMC flag so that it can determine whether or not to try to retrieve the sector logic container in initialize().
The configuration has been updated to ensure that the flag is set correctly.
Also, I removed the readOfflineMuonContainer method of RpcLv1RawDataEfficiency: since this was attempting to retrieve the Run-1 muon container, it was clearly unused.
Mentioning ATLASRECTS-4390 for bookkeeping purposes.


Former-commit-id: f585dc63c7d94ef2bb128bd9e9a16925671cf88c