Added protection for ROS-ROB maps which list ROB IDs which are not available in the used datafiles. The fix is related ATR-17265.
This fix is only relevant for the simulation of ROS hit rates with ROS-ROB
maps, which do not list the same ROB IDs which are also available in the
used datafile.

When running in a partition, only ROB IDs which are listed in the ROS-ROB
map are available for readout. The fix brings the emulation in line with
the behavior of the DCM application by suppressing all ROB IDs from the datafile
which are not listed in the ROS-ROB map. In case no ROS-ROB is provided no
action is done.


Former-commit-id: f58239c0e614245706c9c95d560d3947e9c486fb