Update of online monitoring for 2018 tau triggers
This MR implements the online monitoring (ATR-17829, ATR-17806) for 2018 tau triggers running the RNN identification algorithm (which will supersede the BDT identification, hopefully).
A minor modification was made to TauJetRNN.h in order to be able to retrieve a map that caches all calculated variables used as input to the RNN.
The new variables to be monitored were added to TrigTauRecMerged.cxx, and the corresponding histograms are defined in TrigTauRecMonitoring.py.
I set the 'urgent' label because this MR is needed for the startup release.
Thanks!


Former-commit-id: c30332d6e43c3684867e3f855388a9fa50323716