Three requested JetUncertainties updates
This MR incorporates three requested features.

1. The CalibArea tool property now defaults to CalibArea-02, so users
do not have to specify the property.  This should be kept up to date
in the future so that it evolves to always be the currently
recommended CalibArea version.

2. As discussed on ANALYSISTO-370, the tool now reads a default
flavour composition file and uses that in case the user has
specified nJets behaviour but encounters an event with a different
number of jets than specified.  In this case, the tool should now
fall back on the inclusive "50/50" composition from the default file
while using the user-specified composition files where they have
been provided for a given nJets value.

3. As requested by @cpollard, and also discussed with @kpachal
experimental functionality to support jet-by-jet flavour
uncertainties has been added.  This is NOT for general use, but
rather for investigation into the viability of such an approach.
In this case, quark-jets will have quark-jet uncertainties, and
gluon-jets will have gluon-jet uncertainties, rather than the
current approach of a general response and composition uncertainty
representing the ensemble of input jets and the expected composition.
No config files that use this functionality are yet provided, and
will not be provided generally until this functionality is approved
for general use.

Validation has been performed as follows:

1. Simple validation by checking if it reads the correct path when
the CalibArea has not been set.  This worked as intended.

2. Comparisons of pre-change and post-change with the unknown and
dijet compositions were tested, as those I had available.  Nothing
changed, as intended.  However, I did not test any nJets composition
yet, which is why this MR is currently marked as WIP.  I need input
from the top group on a realistic nJets composition file for me to
use for validation.  I will mention this on ANALYSISTO-370.

3. Only trivial validation has been performed so far, as the exact
approach to follow in jet-by-jet flavour uncertainties is not yet
100% final.  The functionality has been confirmed to not impact
any existing functionality by several tests.  However, actual
functionality has not yet been demonstrated, which is why this
MR is currently marked as WIP.


Former-commit-id: a2fa41d0315356b688a5e064eaa4b5e7aeb71d59