Migrate TgcRawDataMonitoring to use ReadHandles for data retrieval
In addition to the migration, I removed several unused hit container locations from TgcLv1RawDataValAlg.
Also, I added a UseExpressStream property (set to true by default), because it wasn't clear to me whether or not the express stream should always be available.
And in the new framework, it's not possible to check whether or not it's present before deciding whether or not to retrieve the container.
If necessary, this check can be moved to the configuration.


Former-commit-id: 9efa08ba89bedc1b920f4f3bbeb6444201db0156