Removed Extended T/P converters for Track and SegmentCollection
New TLP objects and converters in a new packages (had to be created to
avoid circular dependencies): TrkEventTopLevelCnv with
SegmentCollection_tlp4 and TrackCollectionCnv_tlp6
There are no more Extended T/P converters left and the ability to write
them has been removed. But almost all of the code still remains to be
able to read old data.


Former-commit-id: 376ea4690f2f6d7ec5161c7db134d55c58383a8a