Add additional tau validation jobs for the RTT and ART frameworks
Added two additional tests for the tau chains in ...
```
  Trigger/TrigValidation/TrigInDetValidation/TIDAbuild/TrigInDetValidation_tau_IBL_pu46_NoDoublet.xml
  Trigger/TrigValidation/TrigInDetValidation/TIDAbuild/TrigInDetValidation_tau_IBL_pu46_RoiParam.xml
```
Included changes to the master makefile to include these.

The files
 ```
        Trigger/TrigValidation/TrigInDetValidation/test/test_trigindetvalidation_tau_ibl_pu46_nodoublet.sh
	Trigger/TrigValidation/TrigInDetValidation/test/test_trigindetvalidation_tau_ibl_pu46_roiparam.sh
        Trigger/TrigValidation/TrigInDetValidation/TIDAbuild/TrigInDetValidation_TestConfiguration_main.xml
```
are autogenerated and should not be reviewed. Also included some changes to the
core job options file to allow the additional settings that these jobs use ..
```
  Trigger/TrigValidation/TrigInDetValidation/share/TrigInDetValidation_RTT_Common.py
```
As usual, the statement that non of this code is used at Tier 0 or at P1, these are
just tests for the RTT and ART frameworks, and have no impact on the Tier 0 operation.


Former-commit-id: c301e6acaf177aad5c474d4dca95f42b4d94b752