ATLASJT-363: Fix bug, JobReport doubles dbDataTotal and dbTimeTotal
The implementation of trfJobReport, had assumption of generating one
jobReport per job. This is not valid on grid jobs, since a text report
is automatically generated by default in addition to JSON report.
This caused summing of dbTime and dbTotal, each time a report is
generated. This fix separates logic for report data generation from
that of report output for each format. The data dictionary is generated
only once and reused for each report format. The fix eliminates
the assumption of one report per job.


Former-commit-id: de0586891b691be8dbedfc187dc2be89527bef31