Workaround for the bug in fastjet-contrib ConstituentSubtractor discussed in ATLASG-1417.
There was a seg fault when using ConstituentSubtractorTool.cxx. It happens very rarely
for events in which the particle with minimal rapidity has special value (dependend on the
parameters used in the ConstituentSubtractor). In this commit, a workaround is added,
in which the rapidity of these particles is slighly shifted. It is performed only for the
problematic events. This workaround should not affect the performance of the
ConstituentSubtraction.


Former-commit-id: c5e5f6903eda560f83d14323c5b5b33b14188031