Fix for ATLASRECTS-4422
This fixes the crash reportd in ATLASRECTS-4422 by adding protection against the MdtRawDataValAlg trying to retrieve the LVL1MuonRoIs when they are not available.
Eventually the check for the presence of this collection should probably be migrated to the configuration, but for now the important thing is to get things running again.


Former-commit-id: 9d9eb4ed3ac5728f0c0c0aa5773719a4277ea410