Protection to serve ROBs from the local ROB cache, which are on the ignoreROB list
In case a list of ROBs was defined, which should be ignored for retrieval, it was possible
that these ROBs could still be served from the local cache. This was however only possible
if the local cache was filled before by a retrieval of the complete event data (i.e. event
building). This update introduces a check that ROBs which should be ignored can not be
added anymore to the local cache.


Former-commit-id: e56ab1f145f0add15bfb30bcd10c913378ffaf9a