Move all configuration for ISF_ParticleFinalStateFilter into the CfgGetter method
There was some lingering configuration for `ISF_ParticleFinalStateFilter` still
in `iGeant4.py` which was also problematic as it lead to the creation of a public
version of the AlgTool. This should fix ATLASSIM-3634.
Decided to configure `ISF_ParticleFinalStateFilter` in the same way for `AtlasG4`
and ISF `FullG4` jobs. Previously the ChecGenSimStable and CheckGenInteracting
properties were set to be `True` in `AtlasG4`. NB Did not observe any difference
in output in local `AtlasG4_tf.py` tests.


Former-commit-id: c58c7a7edcbf58a87dba4724ac1fe0c3323cc2a7