infrastructure to write XML geometry with AGDD (ATLASSIM-3636)
This is a first MR to eventually allow the inheriting AGDD tools
to write their XML geometry description into a dedicated output
file. With some code following in a separate MR the corresponding
SQL files will be created to load the XML blob to the DB.

Comments to the shifter:

1/ Please don't suggest to remove the comment in XercesParser.cxx
It took some time to find it in the web and i would like to save
this time for future developers.

2/ Please don't propose to use a smart pointer in XercesParser.cxx
As the pointer is used by the external Xerces software i am not
sure how it would react on this change and don't wanna try out.

3/ Please don't ask if the cout can be avoided. AGDD is a software
written to operate outside the Athena, too. To introduce the Athena
logger would lead to a unnecessary dependency and destroy this
possibility.


Former-commit-id: 07f0f3f4f95f486b7e19878a35d93b81a23fb80c