Patched GeoModelUtilities/GeoAlignmentStore by protecting insertions into position cache maps
It is possible that two different threads ask the same FullPhysical volume for its absolute
positon and this may trigger concurrent attempts to store the position cache into the same
alignment store. This is a normal behavior and should not throw any exception (fixed).
Also the attempts to emplace position caches are now protected by a mutex


Former-commit-id: ce71abcc30aa24cebfe740e5acfa1ee757beb2d2