Fix Coverity Defect 115443 in LArFCALCalcuatorBase.cxx
The `m_OOTcut` variable is already defined as a configurable property
of the `LArCalculatorSvcImp` base class, so it does not need to be
re-defined in the `LArFCALCalcuatorBase` derived class. In addition
rather than hard-coding the value during `LArFCALCalcuatorBase::initialize`
it can be set in the relevant `CfgGetter` methods in `LArG4FCALConfig.py`.
This should fix Coverity Defect 115443.


Former-commit-id: bca611904cd6721b365aa1c2e67a89fb20ce8bb4