AthContainers: Add a notion of atomic access to auxiliary variables.
Allow an aux variable to be marked as atomic.  If so, then we want it to be
accessed only via AtomicDecorator, which will cast the value to a proper
atomic type.

Please not that this does not completely prevent non-atomic access to such
atomic variables.  In particular, the implicit accesses done via the
AuxTypeRegistry operations are not atomic, and it is possible to get
non-atomic access to the variables by requesting the variable vector directly.
Some of these could be fixed, but at a cost of significant extra coding.
Therefore, this facility is not recommended for general use at this point.
Please consult with core software before adding new uses of this.

Added a flag argument to the getAuxID methods in AuxTypeRegistry
Added AtomicDecoration.


Former-commit-id: 5258db66801b9670ce58ca855715fac7a7e72d3d