writing AGDD geometry blob for DB from amdc_simrec (ATLASSIM-3636)
This merge request is a follow-up on !10413 and uses the new
infrastructure to write data base blobs with AGDD. In this
merge request the MuonAGDDTool was extended in order to write
the blob from AGDD part of the amdc_simrec file parsed by
AmdcAth. In principle this is *not* needed and there is no
dependecy on AMDB. The MuonAGDDTool could also be configured
to write the blob from any XML file. To not modify the workflow
too much this isn't done (yet). Previously the blob was written
by AmdcDbSvcMakerFromAmdc and the corresponding PreSql file for
uploading was done by AmdcDb2Sql. Now both is done within the
MuonAGDDTool. To streamline the creation of the geometry
a set of python files was added. They allow for the creation
AMDB and AGDD Sql files and tables/blobs separately in a
very fast job.

Comment to shifter: There are unavoidable couts as in these
places the ATH_MSG macros are not available. Please refrain
commenting on that.


Former-commit-id: af907a99dfbfe4d57d634670560f9c1c12846493