SUSY3 update for 2018 open-ended production (ATLSUSYDPD-1310)
This MR extends the SUSY3 skimming requirement on offline taus.
In 2017, we were skimming on loose BDT taus.
To get prepared for the 2018 RNN ID, I made an OR between BDT loose and RNN loose.
There is a detection of the offline reconstruction release, and if it's above 21.0.63, RNN is enabled in the skimming.
This was tested on old 2017 data (which have no RNN) and new data (2017 reprocessed data), and old MC.
We don't yet have "new MC" reconstructed with 21.0 that has RNN.
We have some MC trigger-specific samples reprocessed with 21.5.4 (escaping the 21.0.X detection), but SUSY3 is not meant to be run on this kind of samples.

The RNN WPs are not defined yet centrally in DFTau, so I have to cut on the RNN flattened score, the target effciencies being defined here:
https://gitlab.cern.ch/atlas/athena/blob/21.0/Reconstruction/tauRec/python/TauAlgorithmsHolder.py#L892

For MC, we can expect a size increase starting from MC16e, but everything up to MC16d won't have RNN tau ID.
For data, based on a trigger+reco reprocessing sample, the size increase is X%.

Cheers,
Bertrand


Former-commit-id: 38fa22341c3ee2980e7dc640014e4e94a9cad156