adjustment for Run-3 geometries
The table AtlasCommon attached to each layout features a column
to determine the run period. Up to now only the strings "RUN1",
"RUN2" and "RUN4" exist on the database. "RUN3" isn't used there
yet, but will be part of the future NSW layouts. The inner
detector is for now the main client of this flag and treated
"RUN3" like "RUN4", i.e. with ITk upgrade. "RUN3" will however
certainly not have the ITk. Therefore the code is adusted to treat
"RUN3" similar to "RUN2" rather than "RUN4". Once the NSW layout
is in place "RUN3" will be actively used by the muon system.


Former-commit-id: fa0414e0c91039114e3a8ddc6d493ce4580dc4d7