introduce "RUN3" string as alternative to NSW simflag
To simulate the NSW it is currently needed to set the simflag
"SimulateNewSmallWheel" to be true. This is a manual action.
In the future all jobs (including simulation jobs) should
get this configuration automatically form the geometry, more
precisely from the run period encoded in the AtlasCommon
table. A first layout with this table filled properly is
now in place. Therefore this MR introduces the "RUN3" string
as alternative to the simflag. Once it is more established
the simflag can be removed.


Former-commit-id: 1576d08a1650ff449ecdc32e2485df79abde2b3c