mods of NSWAGDDTool to write blob for NSW layouts (ATLASSIM-3636)
This merge request includes all coded needed in order to produce
and upload the NSW geometry AGDD XML file to the database. The
XercesParser was adjusted to remove all comments from the XML
and adjustments in the WritePREsqlFile method of the NSWAGDDTool
and MuonAGDDTool will remove blank lines in the DB blob. The
MuonAGDDToolHelper was modified to be used by both AGDDTools to
retrieve the blob from the database. The NSWAGDDTool is extended
to be capable of using this functionality of the helper. The
AmdcAth_GeometryTasks.py script is modified and now aware of the
new scripts NSWAGDD_BlobProduction.py and NSWAGDD_BlobCheck.py
used to upload and check the database tables. The script to
generate the SQL files (ProduceSql.sh) was changed to recognize
the NSWD PreSql properly. Finally a new layout stations.v2.00.xml
is uploaded. It includes the current best description of the
NSW including dead material (like JD and Hub). It is not yet
linked as default in NSW_StationDescription.xml to avoid
complications, but used in ATLAS-R3-2012-00-00-00.


Former-commit-id: 304fc2e503a53daf14768b9694b9728470cf9543