use module hash to identify NSW PrepData and RawData
Up to now the detectorelement hash is used to identify the NSW
PrepData and RawData. It is more advicable to use module hashes.
This merge request applies the according modifications. As the
data collections aren't really used yet it's possible to just
change the converters rather than doing a schema evolution.


Former-commit-id: 8f979d7fe493d2ad0d74d58399b48509c7125b3c