Patching truth decay collection maker to follow passing children
This solves an open issue with our truth decay collections.  In the case
of some generators, multiple particles of interest are written into a
chain.  At the moment, the first one is encountered and added to the
collection (chain).  The code goes down the tree by N generations
according to the user settings.  In order to not keep multiple copies,
subsequent particles that have already been added to the collection are
not re-recorded.  Of course, this leads to a situation where the first N
generations are kept for the first particle, the first N-1 generations
are kept for the second, and so on -- even if the second particle would
have passed the cuts and should have been saved with N generations
itself.

In other words, for a 1-generation collection of ttbar decays, the first
top gets its decay products saved, but the second top (if it comes after
the first in the tree) does not.

Patch solves the problem and refactors the "pass" code so that it can be
reused elsewhere.


Former-commit-id: 71a757d0cd49b25c3b9b2b278316830d0ba00267