NSW geometry auto-config for digitization and reconstruction
The NSWAGDDTool is building the NSW geometry. In simulation there
was the simulation flag "SimulateNewSmallWheel" while no such
flag was in place for digitization/reconstrution and postIncludes
had to be used. With the availability of "RUN3" in the AtlasCommon
table it is now possible to use this string for digitization and
reconstruction jobs to trigger the building of the NSW geometry.
This merge request is taking care of that.


Former-commit-id: a9b81e65b9569161bb639c58e646e08890c3b129