Removing FATAL crash from InDetForwardTrackParticleThinning
The InDetForwardTrackParticleThinning algorithm returns a FATAL error if either the InDetForwardTrackParticle or Muons container is absent. Since this causes some TCTs to fail, the FATAL error is replaced with a WARNING, and the FAILURE is replaced with SUCCESS - in which case, the algorithm does nothing.

This should not affect any output and consequently does not break frozen Tier-0


Former-commit-id: 3d6577131f0494945b972bc4a2b3ea7394fff221