AthenaPoolUtilities, RecAthenaPool: Allow skipping tests if required libraries aren't present.
Synchonize with 21.2:

Some *AthenaPool tests are failing in AthDerivation because they require
libraries that aren't present in the build.  For example, RecAthenaPool
is failing because JetEventAthenaPool isn't included.  In this case,
we end up having to add a fair number of libraries to the build
in order to get it to work --- probably more than we'd like to.

So extend the cmake_parse_arguments macro to take a REQUIRED_LIBARARIES
argument.  If any of the libraries listed here are not found
on LD_LIBRARY_PATH, then the test is skipped.

Use this in RecAthenaPool.


Former-commit-id: 71ff378c73757fe155ff4dd765c5c36638d3ea68