Correctly set the sampling for E4' chambers in ISF_Analysis (ATLASSIM-3601)
It was observed that some Tile Identifiers were returning sampling(layer) values
of 23 (which corresponded to an FCAL layer) this was causing problems in the
FastCaloSim V2 parameterization. Experts confirmed that the E4' Identifiers do not
fit with the usual Tile Identifier scheme and as such the sampling should be
hard-coded as `TileGap3` for such Identifiers. This should fix ATLASSIM-3601.


Former-commit-id: ffa67cf81f7a693062dbe15a9c13952b8fa0732b