Support for 0-track taus in TauWPDecorator
This MR is the first of a series for tau trigger.
A big MR !10692 is going to be truncated into a few smaller independent MRs, to ease validation.
This one brings improvement to the TauWPDecorator:
- support for 0p (to be used for tau trigger, a priori not for offline)
- some code cleanup
- drop accessors and use instead tau EDM variables: we no longer need to run TauIDVarCalculator just to decorate a handful of variables for the RNN flattening (ATLTAU-1505).

I've set the 'urgent' label because other MRs in preparation depend on this one.

@fscutti , @bailey , @cdeutsch : for information, I did not drop the decorations from TauIDVarCalculator, because the old Run-1 tau trigger BDT might still depend on these.
Since Run1 tau triggers are still defined in TauDef.py, although they may have been completely phased from the trigger menu out by now, it may be wise to keep TauIDVarCalculator untouched until the end of this year.

WARNING: in order to keep the previous behaviour, if no Root file is provided for 0p, 0p will use the 3p file (the code looked like: if nTracks==1 use 1p, else use 3p).

Cheers,
Bertrand


Former-commit-id: 444863678395f7428f17269a7821ea9efc564921