fixing wrong ordering of config in sim jobs (!10664)
MR !10664 introduced the CommonGeometryFlags in configuration file
G4Atlas.flat.configuration.py . Unfortunately these flags got used
before the geometry was set up entirely, and therefore the flags
were initialized with default values rather than the ones of the
layout. In this MR the problem is resolved by moving the proper
setting of the layout before the initialization of the flags.


Former-commit-id: 4730c372e3709756c65c79617979c8a61368f069