fixing get_files script to retrieve locally installed files
If code is checkout locally and atlas_install_scripts is added (or
extended) in the CMakeLists.txt file in order to make (more)
scripts available, it does work in the sense that the scripts can
be executed. However, using the "get_files" script to download
or better copy the file fails and reports the file cannot be
found. If job options, data files and xmls are installed by in
local modifcation of the CMakeLists.txt, those files can be found.
The script even recognizes installed scripts which are already part
of the build, i.e. centrally available. Only local script files
cannot be found even though they can be exectued.

This MR should prevent developers to get (more) insane and frantic
by adjusting the path variable in get_files ........... o_O


Former-commit-id: 215448915989cc6f4a967942d6f7a059892e2940