KLFitter ttZ: use calo eta for electrons, not track eta
This fixes a bug with the interface to the ttZ likelihood of KLFitter: the
corresponding methods in KLFitterTool (TopEventReconstructionTools) add
electrons from the xAOD containers to the KLFitter particle collection, but used
the track eta, instead of the calo eta. Since this propagates to the decision,
which eta bin to use for the transfer function parameterisation, the code fails
when the track eta falls into the crack region. Those cases are rare, but
possible, since the crack region is defined based on the eta of the calo cells.

This is now fixed by using the calo eta instead of track eta for electrons.


Former-commit-id: aea9c050cdd8515927907f9c0f8708bc152b7416