updating NSW simulation ART tests
The new style of AGDD configuration using the database made the
ART jobs of the NSW crash. There is now an update to have a job
running the old configuration (w/o crashing) and an additional
job which is based on the new configuration. Once the new
configuration is fully established the test of the old config
can be removed.


Former-commit-id: b56ce12e75a1900ea218316042aca256503995f3