reducing verbosity of NSW readout elements
The NSW readout elements produce a lot of verbose log which was
used frequently for debugging (and may be needed again). This MR
sends most of the info to debug. There are a few actual couts which
were not remove, but just commented out. It would be prefered to
keep them because locally the developers may still need them a lot.
The other couts/cerrs cannot be replaced by the Athena msg macro.
Reducing the verbosity changed a number of unit test results. The
reference files are updated accordingly. There are four unit test
in MuonEventTPCnv expected to fail. It's known and therefore no
show stopper to proceed with the MR.


Former-commit-id: 14d40e8ee69d76a9572c31f38ed4a18f7dc3f158