cut backward compatibility for layouts < R.04.05 (ATLASRECTS-2134)
Global geometry tags supported by releases 21+ do only include
MuonSpectrometer layouts of type R.06 or later. Therefore the
support for layouts of type R.04.05 and earlier can be cut. This
MR cleans up MuonGeoModel for if-statements which act on layouts
with digit combinations 04.05=405 in the name or earlier. To
check that there is no difference introduced acmd.py diff-root
was performed no hit files of 2000 single muons events produced
with ATLAS-R2-2015-03-01-00 .


Former-commit-id: 282257d501c041299d42c98b875106353a84c230