removal of Muonboy MuonDriftCircleErrorStrategy
Muonboy had its own strategy how to account for RT and alignment
uncertainties in single hit errors. The uncertainties were derived
by experts in run-1 and hardcoded in the Muonboy code. When it
came to the unification of muon reco algorithms, the Muonboy hit
errors got hardcoded as an option to study what performs best
in competition with the other uncertainties available. As it
was found that Muonboy uncertainties should not be used in the
unified chain, the corresponding code is not longer used. In order
to clean up, this MR removes the option and related methods (today
there is likely anyway nobody left who understands them). The MR
does not violate the frozen tier0 policiy, i.e. it's rather certain
that indeed nobody uses them anymore by chance/mistake. There was
also some other cleanup performed.


Former-commit-id: 03f14e677d9f1a93006786d14e404ff85c2c4178