Cleaning up CopyTruthJetParticles for ATLASG-1375
An issue was reported in ATLASG-1375, where there seems to be an
inconsistency between the leptons used for the identification of FSR
photons and the leptons identified as prompt.  I'm suspicious that this
is coming from the inconsistent definitions that were used in two places
in the code, and this now makes sure that we re-use the work that we've
done (it should also speed things up somewhat).


Former-commit-id: 09c67c82af790815e12f1b8c284ea8ba6d63c8f8