PyUtils: Fix length inconsistency in SetSize call.
The pythonization of TFile::root was calling the SetSize method
on a PyROOT buffer object.  The argument that SetSize takes is the
number of elements in the buffer, where the size of each element
depends on how the buffer was created.  Here, however, we were
passing to SetSize a number of bytes.

The buffer elements we were seeing were void*, which have size 8.
So the buffer size was getting misset by a factor of 8.
This was leading to a crash in the input file peeker.

This crash was however only observed in the muRunData test of TrigUpgradeTest,
and only when it was run with --stdcmalloc and MALLOC_CHECK_=3.
It's a bit a mystery why this apparently long-standing bug hasn't 
caused problems before.



Former-commit-id: b3df5784871ab170818469982bb385a49fe218ba